Abstract nuclei detection (#1045)
* added abstract library for nuclei detection
