+++ histomicstk/segmentation/nuclear/__init__.py
@@ -4,6 +4,7 @@ segmenting nuclei from histopathology images.
 """
 
 from .detect_nuclei_kofahi import detect_nuclei_kofahi
+from .detect_tile_nuclei import detect_tile_nuclei
 from .gaussian_voting import gaussian_voting
 from .gvf_tracking import gvf_tracking
 from .max_clustering import max_clustering
@@ -14,5 +15,6 @@ __all__ = (
     'gaussian_voting',
     'gvf_tracking',
     'max_clustering',
-    'min_model'
+    'min_model',
+    'detect_tile_nuclei'
 )
